requirements.txt
streamlit 
Pillow 
opencv-python 
torch 
torchvision 
numpy
matplotlib
